refactor: 일단 공통 UI 컴포넌트 분리만
