+++ src/6_shared/ui/data/Table.tsx
@@ -1,0 +1,99 @@
+import { forwardRef } from 'react';
+
+/**
+ * 테이블 컴포넌트 계열
+ * - 데이터를 구조화된 표 형태로 표시하는 컴포넌트들
+ * - Table, TableHeader, TableBody, TableRow, TableHead, TableCell로 구성
+ * - 반응형 스크롤 지원과 일관된 스타일링 제공
+ */
+
+interface TableProps extends React.HTMLAttributes<HTMLTableElement> {
+  className?: string;
+}
+
+export const Table = forwardRef<HTMLTableElement, TableProps>(
+  ({ className, ...props }, ref) => (
+    <div className='w-full overflow-auto'>
+      <table
+        ref={ref}
+        className={`table-fixed w-full caption-bottom text-sm ${className}`}
+        {...props}
+      />
+    </div>
+  )
+);
+Table.displayName = 'Table';
+
+interface TableHeaderProps
+  extends React.HTMLAttributes<HTMLTableSectionElement> {
+  className?: string;
+}
+
+export const TableHeader = forwardRef<
+  HTMLTableSectionElement,
+  TableHeaderProps
+>(({ className, ...props }, ref) => (
+  <thead ref={ref} className={`[&_tr]:border-b ${className}`} {...props} />
+));
+TableHeader.displayName = 'TableHeader';
+
+interface TableBodyProps extends React.HTMLAttributes<HTMLTableSectionElement> {
+  className?: string;
+}
+
+export const TableBody = forwardRef<HTMLTableSectionElement, TableBodyProps>(
+  ({ className, ...props }, ref) => (
+    <tbody
+      ref={ref}
+      className={`[&_tr:last-child]:border-0 ${className}`}
+      {...props}
+    />
+  )
+);
+TableBody.displayName = 'TableBody';
+
+interface TableRowProps extends React.HTMLAttributes<HTMLTableRowElement> {
+  className?: string;
+}
+
+export const TableRow = forwardRef<HTMLTableRowElement, TableRowProps>(
+  ({ className, ...props }, ref) => (
+    <tr
+      ref={ref}
+      className={`border-b transition-colors hover:bg-muted/50 data-[state=selected]:bg-muted h-14 ${className}`}
+      {...props}
+    />
+  )
+);
+TableRow.displayName = 'TableRow';
+
+interface TableHeadProps
+  extends React.HTMLAttributes<HTMLTableHeaderCellElement> {
+  className?: string;
+}
+
+export const TableHead = forwardRef<HTMLTableHeaderCellElement, TableHeadProps>(
+  ({ className, ...props }, ref) => (
+    <th
+      ref={ref}
+      className={`h-10 px-2 text-left align-middle font-medium text-muted-foreground [&:has([role=checkbox])]:pr-0 ${className}`}
+      {...props}
+    />
+  )
+);
+TableHead.displayName = 'TableHead';
+
+interface TableCellProps extends React.HTMLAttributes<HTMLTableCellElement> {
+  className?: string;
+}
+
+export const TableCell = forwardRef<HTMLTableCellElement, TableCellProps>(
+  ({ className, ...props }, ref) => (
+    <td
+      ref={ref}
+      className={`p-2 align-middle [&:has([role=checkbox])]:pr-0 ${className}`}
+      {...props}
+    />
+  )
+);
+TableCell.displayName = 'TableCell';
