+++ src/6_shared/ui/forms/Input.tsx
@@ -1,0 +1,25 @@
+import { forwardRef } from 'react';
+
+/**
+ * 입력 컴포넌트
+ * - 기본 HTML input 요소를 확장한 커스텀 입력 컴포넌트
+ * - 일관된 스타일링과 접근성 기능 제공
+ * - forwardRef를 통한 ref 전달로 포커스 관리 지원
+ */
+
+interface InputProps extends React.InputHTMLAttributes<HTMLInputElement> {
+  className?: string;
+}
+
+export const Input = forwardRef<HTMLInputElement, InputProps>(
+  ({ className, type, ...props }, ref) => {
+    return (
+      <input
+        type={type}
+        className={`flex h-10 w-full rounded-md border border-input bg-white px-3 py-2 text-sm ring-offset-background file:border-0 file:bg-transparent file:text-sm file:font-medium placeholder:text-muted-foreground focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:cursor-not-allowed disabled:opacity-50 ${className}`}
+        ref={ref}
+        {...props}
+      />
+    );
+  }
+);
