+++ src/6_shared/ui/forms/Textarea.tsx
@@ -1,0 +1,26 @@
+import { forwardRef } from 'react';
+
+/**
+ * 텍스트 영역 컴포넌트
+ * - 여러 줄 텍스트 입력을 위한 textarea 컴포넌트
+ * - Input 컴포넌트와 일관된 스타일링 유지
+ * - 최소 높이 설정으로 사용성 향상
+ */
+
+interface TextareaProps
+  extends React.TextareaHTMLAttributes<HTMLTextAreaElement> {
+  className?: string;
+}
+
+export const Textarea = forwardRef<HTMLTextAreaElement, TextareaProps>(
+  ({ className, ...props }, ref) => {
+    return (
+      <textarea
+        className={`flex min-h-[150px] w-full rounded-md border border-input bg-white px-3 py-2 text-sm ring-offset-background placeholder:text-muted-foreground focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:cursor-not-allowed disabled:opacity-50 ${className}`}
+        ref={ref}
+        {...props}
+      />
+    );
+  }
+);
+Textarea.displayName = 'Textarea';
