+++ src/6_shared/ui/index.ts
@@ -1,0 +1,16 @@
+// controls
+export * from './controls/Button';
+
+// data
+export * from './data/Table';
+
+// forms
+export * from './forms/Input';
+export * from './forms/Select';
+export * from './forms/Textarea';
+
+// layout
+export * from './layout/Card';
+
+// overlays
+export * from './overlays/Dialog';
